Reviewed-on: https://codeberg.org/xwzCAD/FreeCAD/pulls/195

+++ src/Mod/ReverseEngineering/App/AppReverseEngineering.cpp
@@ -403,7 +403,7 @@ private:
                 }
             }
 
-            TColgp_Array1OfPnt clPoints(0, pts.size()-1);
+            TColgp_Array1OfPnt clPoints(0, static_cast<int>(pts.size() - 1));
             if (clPoints.Length() < uPoles * vPoles) {
                 throw Py::ValueError("Too less data points for the specified number of poles");
             }
